+++ client_src/js/filamentManagement.js
@@ -618,8 +618,7 @@ async function init() {
   historyStatistics = await historyStatistics.json();
   let usageByDay = historyStatistics.history.totalByDay;
   let usageOverTime = historyStatistics.history.usageOverTime;
-  console.log(usageOverTime);
-  console.log(usageByDay);
+
   const usageOverTimeOptions = {
     chart: {
       type: "bar",
@@ -712,11 +711,6 @@ async function init() {
       animations: {
         enabled: true,
       },
-      plotOptions: {
-        bar: {
-          horizontal: false,
-        },
-      },
       toolbar: {
         show: false,
       },
@@ -724,27 +718,6 @@ async function init() {
         enabled: false,
       },
       background: "#303030",
-    },
-    // fill: {
-    //   type: "gradient",
-    //   gradient: {
-    //     shade: "dark",
-    //     gradientToColors: ["#FDD835"],
-    //     shadeIntensity: 1,
-    //     type: "horizontal",
-    //     opacityFrom: 1,
-    //     opacityTo: 1,
-    //     stops: [0, 100, 100, 100],
-    //   },
-    // },
-    markers: {
-      size: 4,
-      colors: ["#FFA41B"],
-      strokeColors: "#fff",
-      strokeWidth: 2,
-      hover: {
-        size: 7,
-      },
     },
     dataLabels: {
       enabled: false,